% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extmonth.R
\name{extmonth}
\alias{extmonth}
\title{Extraction Function (Month)}
\usage{
extmonth(data, year, degree, month)
}
\arguments{
\item{data}{Name of the input data. The format of the data follows the output value of the 'extncdf4' function in the same package.}

\item{year}{Years to be extracted}

\item{degree}{Interval between measurements of NCEP data (e.g. 2.5 x 2.5 -> 2.5)}

\item{month}{Month to be extracted (e.g. Jan -> 1, Aug -> 8)}
}
\value{
A matrix of the infile
}
\description{
This function works based on the data extracted by the extncdf4 function.
Extracts a specific month value for a specific year.
}
\examples{
Jan_Uwind = extmonth(Uwind,1985:2020,2.5,1)
}
