% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extncdf4.R
\name{extncdf4}
\alias{extncdf4}
\title{Extraction Function}
\usage{
extncdf4(filename, level, dsy, year)
}
\arguments{
\item{filename}{Name of the input ncdf4 file (e.g. 'uwnd.mon.mean.nc')}

\item{level}{Barometric pressure to be output (e.g. 925hPa -> 925)}

\item{dsy}{Start year of original data (e.g. Jan,1948 -> 1948)}

\item{year}{Range to be output (e.g. 1985 to 2020 -> 1985:2020)}
}
\value{
A matrix of the infile
}
\description{
This function loads the ncdf4 file of NCEP as a matrix.
}
\examples{
Uwind = extncdf4('uwnd.mon.mean.nc',850,1948,1985:2020)
}
