% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extseason.R
\name{extseason}
\alias{extseason}
\title{Extraction Function (Seasonal)}
\usage{
extseason(data, year, degree, start, end)
}
\arguments{
\item{data}{Name of the input data. The format of the data follows the output value of the 'extncdf4' function in the same package.}

\item{year}{Years to be extracted}

\item{degree}{Interval between measurements of NCEP data (e.g. 2.5 x 2.5 -> 2.5)}

\item{start}{Start of season to be extracted (e.g. June to November -> 6)}

\item{end}{End of season to be extracted (e.g. June to November -> 11)}
}
\value{
A matrix of the infile
}
\description{
This function works based on the data extracted by the extncdf4 function.
Extracts a seasonal value for a specific year.
}
\examples{
JJASON_Uwind = extmonth(Uwind,1985:2020,2.5,6,11)
}
