% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial.R
\name{partial}
\alias{partial}
\title{Distance Calculating Function}
\usage{
partial(data, degree)
}
\arguments{
\item{data}{Name of the input data. The format of the data follows the output value of Extraction function ('extmonth' or 'extseason') in the same package.}

\item{degree}{Interval between measurements of NCEP data (e.g. 2.5 x 2.5 -> 2.5)}
}
\value{
A matrix of the infile
}
\description{
This function calculates the distance of coordinates on the ellipsoid.
}
\examples{
distance = partial(JJASON_Uwind,2.5)
}
