% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windcomp.R
\name{windcomp}
\alias{windcomp}
\title{Wind Component Calculating Function}
\usage{
windcomp(uwnd, vwnd, degree, year, component)
}
\arguments{
\item{uwnd}{Name of the meridional wind. The format of the data follows the output value of Extraction function ('extmonth' or 'extseason') in the same package.}

\item{vwnd}{Name of the zonal wind. The format of the data follows the output value of Extraction function ('extmonth' or 'extseason') in the same package.}

\item{degree}{Interval between measurements of NCEP data (e.g. 2.5 x 2.5 -> 2.5)}

\item{year}{Range to be analysis (e.g. 1985 to 2020 -> 1985:2020)}

\item{component}{The wind component to be calculated. This option has a choice of 'rotational' and 'divergent'.}
}
\value{
A matrix of the infile
}
\description{
This function calculates the Component of Wind.
Winds can be decomposed into rotational and divergent components. These are outputted as vorticity and divergence, respectively.
}
\examples{
vorticity = windcomp(JJASON_Uwind,2.5,1985:2020,'rotational')
}
